Reviewed and updated UX on Windows
