+++ internal/fileutils/fileutils_lin_mac_test.go
@@ -27,26 +27,5 @@ func TestSymlink(t *testing.T) {
 	assert.False(t, IsSymlink(target), "expected no symlink")
 }
 
-func TestIsWritableFile(t *testing.T) {
-	file, err := WriteTempFile(
-		"", t.Name(), []byte("Some data"), 0777,
-	)
-	if err != nil {
-		t.Error(err)
-	}
-
-	if IsWritable(file) != true {
-		t.Fatal("File should be writable")
-	}
-
-	err = os.Chmod(file, 0444)
-	if err != nil {
-		t.Error(err)
-	}
-
-	if IsWritable(file) != false {
-		t.Fatal("File should no longer be writable")
-	}
-}
