+++ internal/offinstall/storage_windows.go
@@ -5,7 +5,7 @@ import (
 	"path/filepath"
 )
 
-func DefaultInstallPath() (string, error) {
+func DefaultInstallParentDir() (string, error) {
 	// There is no system install path for Windows
 	return filepath.Join(os.Getenv("USERPROFILE"), "AppData", "Local", "Programs"), nil
 }
